Fixed issues mentioned in the reviews for PR#2 and other bugs
The given code did not compile at all (probably due to the changes of the column interface).  
I tried to fix all the issues within the code and the comments from the reviews.
_YET_, I could not build the test for avx2 since there is no imask_population_count for avx2. We need to add this to the TSL.
Furthermore, we need to get rid of the tools/tvlgen folder. The .gitmodules is correct but when I clone the repository using --recurse-submodules the TSL files are placed into tools/tvlgen instead of tools/tslgen which produces an error while building (since the CMakeLists assumes, that all the files reside in tools/tslgen). I have no idea how to fix this.
Unfortunately, the meta questions are not answered after 3(!!!) weeks. 
_PLEASE_ answer them as soon as possible.